requirements.txt
# Core GUI
PySide6==6.7.0
PySide6-Addons==6.7.0
PySide6-Essentials==6.7.0

# Scientific Computing
pyscf==2.9.0
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
h5py==3.11.0
rdkit==2024.03.3

# Visualization
matplotlib==3.9.0
vtk==9.3.0
pyqtgraph==0.13.7

# Database
sqlalchemy==2.0.30
alembic==1.13.1
mysql-connector-python==9.3.0

# Utilities
pyyaml==6.0.1
loguru==0.7.2
python-dotenv==1.0.1

# Testing
pytest==8.2.2
pytest-qt==4.4.0
pytest-cov==5.0.0

# Development
black==24.4.2
flake8==7.0.0
mypy==1.10.0

# Packaging
pyinstaller==6.8.0

# GPU support (optional)
gpu4pyscf-cuda12x==1.4.0

# MCP Server (optional)
mcp==1.0.0
fastmcp==0.14.0